# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Settly Core API",
    "version": "1.0.0",
    "description": "Authoritative HTTP API contract for the Settly real estate platform. Implements RFC 9457 error taxonomy, strict validation, and bilingual domain entities."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "settly_session",
        "description": "Better Auth HTTP-only secure session cookie"
      }
    },
    "schemas": {
      "BaseProblem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Relative URI identifier of the specific problem type (machine-readable code)",
            "example": "/errors/validation-failed"
          },
          "title": {
            "type": "string",
            "description": "Short human-readable developer summary of the problem type",
            "example": "Validation Failed"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server",
            "example": 422
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence of the problem",
            "example": "One or more request parameters failed validation schema checks."
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "example": "/api/v1/areas?level=INVALID"
          },
          "requestId": {
            "type": "string",
            "description": "Authoritative request correlation ID (UUIDv7) for telemetry tracing",
            "example": "0191eb45-8f67-73d8-9db8-bc234a9e51c8"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Structured machine-readable parameters for localized client error rendering"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "requestId"
        ]
      },
      "ValidationProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/validation-failed"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  422
                ]
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string",
                      "example": "level"
                    },
                    "code": {
                      "type": "string",
                      "example": "invalid_enum_value"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid enum value"
                    },
                    "params": {
                      "type": "object",
                      "additionalProperties": {
                        "nullable": true
                      }
                    }
                  },
                  "required": [
                    "path",
                    "code"
                  ]
                },
                "description": "Array of validation issues per field"
              }
            },
            "required": [
              "errors"
            ]
          }
        ]
      },
      "NotFoundProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/not-found"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  404
                ]
              }
            }
          }
        ]
      },
      "ForbiddenProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/forbidden"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  403
                ]
              }
            }
          }
        ]
      },
      "UnauthenticatedProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/unauthenticated"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  401
                ]
              }
            }
          }
        ]
      },
...
      "AgentVerificationInput": {
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean",
            "description": "Verification decision: true to approve, false to revoke/reject",
            "example": true
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "description": "Administrative rationale or review notes",
            "example": "License verified against Egyptian Real Estate Authority registry."
          }
        },
        "required": [
          "verified"
        ]
      },
      "AgentListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Agent profile UUIDv7"
          },
          "userId": {
            "type": "string",
            "description": "User ID FK"
          },
          "licenseNumber": {
            "type": "string"
          },
          "brokerageName": {
            "type": "string"
          },
          "bioEn": {
            "type": "string",
            "nullable": true
          },
          "bioAr": {
            "type": "string",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean"
          },
          "verifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "image": {
                "type": "string",
                "nullable": true,
                "format": "uri"
              }
            },
            "required": [
              "id",
              "name",
              "email"
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "licenseNumber",
          "brokerageName",
          "isVerified",
          "createdAt",
          "user"
        ]
      },
      "AgentListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentListItem"
            },
            "description": "Array of agent profiles"
          }
        },
        "required": [
          "items"
        ]
      },
...
      "PropertyImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "0191eb45-8f67-73d8-9db8-bc234a9e51c8"
          },
          "propertyId": {
            "type": "string",
            "format": "uuid",
            "example": "0191eb45-8f67-73d8-9db8-bc234a9e51c9"
          },
          "cloudinaryPublicId": {
            "type": "string",
            "example": "settly/properties/villa-1"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://res.cloudinary.com/dmzcnnxzp/image/upload/v1/settly/properties/villa-1.jpg"
          },
          "captionEn": {
            "type": "string",
            "nullable": true
          },
          "captionAr": {
            "type": "string",
            "nullable": true
          },
          "isCover": {
            "type": "boolean",
            "default": false
          },
          "order": {
            "type": "integer",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "propertyId",
          "cloudinaryPublicId",
          "url",
          "createdAt"
        ]
      },
...
      "AuthorizeUploadResponse": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string",
            "enum": [
              "cloudinary"
            ]
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "constraints": {
            "type": "object",
            "properties": {
              "maxBytes": {
                "type": "integer"
              },
              "allowedTypes": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "maxBytes",
              "allowedTypes"
            ]
          }
        },
        "required": [
          "assetId",
          "provider",
          "uploadUrl",
          "params",
          "constraints"
        ]
      },
      "AuthorizeUpload": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "property_image",
              "document",
              "avatar"
            ],
            "default": "property_image"
          },
          "propertyId": {
            "type": "string",
            "format": "uuid"
          },
          "contentType": {
            "type": "string",
            "pattern": "^image\\/(jpeg|png|webp|avif)$"
          },
          "byteSize": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 15728640
          },
          "filename": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "contentType",
          "byteSize",
          "filename"
        ]
      },
      "CompleteUploadResponse": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "PROCESSING",
              "FAILED"
            ]
          },
          "image": {
            "$ref": "#/components/schemas/PropertyImage"
          }
        },
        "required": [
          "assetId",
          "status"
        ]
      },
      "CompleteUpload": {
        "type": "object",
        "properties": {
          "providerRef": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "cloudinaryPublicId": {
            "type": "string"
          },
          "captionEn": {
            "type": "string"
          },
          "captionAr": {
            "type": "string"
          },
          "isCover": {
            "type": "boolean"
          }
        },
        "required": [
          "providerRef",
          "url",
          "cloudinaryPublicId"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/uploads/authorize": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "summary": "Authorize direct media upload",
        "description": "Issues signed Cloudinary upload parameters for direct client upload per STORAGE.md Section 3.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizeUpload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed upload instructions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeUploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{assetId}/complete": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "summary": "Finalize direct upload",
        "description": "Binds uploaded provider asset to property record per STORAGE.md Section 3.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "assetId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "propertyId",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUpload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload finalized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/agents": {
      "get": {
        "summary": "List agent verification queue (Admin only)",
        "description": "Returns registered agent profiles for moderation and verification review.",
        "tags": [
          "Admin",
          "Identity"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "verified",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of agent profiles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (Admin role required)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/agents/{id}": {
      "get": {
        "summary": "Get agent profile details for review (Admin only)",
        "tags": [
          "Admin",
          "Identity"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent profile details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "404": {
            "description": "Agent profile not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/agents/{id}/verify": {
      "post": {
        "summary": "Approve or revoke agent verification (Admin only)",
        "description": "Updates agent verification status and records an immutable AuditLog entry.",
        "tags": [
          "Admin",
          "Identity"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentVerificationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent verification updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "404": {
            "description": "Agent profile not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
...
  }
}